// packages/pet-overlay/src/pets/renderer-catalog.generated.ts
export const PETS = [
  {
    key: 'mole_digger',
    petId: '003',
    slug: 'mole_digger',
    grade: 'COMMON',
    name: '두더지',
  },
  {
    key: 'sprout_treant',
    petId: '004',
    slug: 'sprout_treant',
    grade: 'COMMON',
    name: '새싹나무',
  },
  {
    key: 'acorn_squirrel',
    petId: '001',
    slug: 'acorn_squirrel',
    grade: 'EPIC',
    name: '도토리다람쥐',
  },
  {
    key: 'star_wizard',
    petId: '006',
    slug: 'star_wizard',
    grade: 'EPIC',
    name: '별빛마법사',
  },
  {
    key: 'cheek_hamster',
    petId: '005',
    slug: 'cheek_hamster',
    grade: 'RARE',
    name: '볼주머니햄',
  },
  {
    key: 'midnight_zebra',
    petId: '002',
    slug: 'midnight_zebra',
    grade: 'RARE',
    name: '미드나잇얼룩말',
  },
] as const;